sec+robustness: close audit findings across PSP, auth, idempotency, SSRF, locking
Sweeping audit-driven hardening across the codebase. Every finding from the
Go-skill audit pass is addressed in this single PR.

Critical (money/auth risk):
- PSP webhook signature verification. Paystack now requires HMAC-SHA512 of the
  raw body against the merchant secret (X-Paystack-Signature); Checkout.com
  requires HMAC-SHA256 against CHECKOUT_WEBHOOK_SECRET (Cko-Signature). Both
  parsers fail closed when no secret is configured. Compares are constant-time.
  The HTTP webhook handler reads the per-PSP header and threads it through
  port.PaymentWebhookPayload.Signature.
- API key plaintext storage replaced with HMAC-SHA256(API_KEY_PEPPER, rawKey).
  Schema: api_keys.key -> api_keys.key_hash (unique). Domain ApiKey carries a
  transient RawKey field surfaced to the caller exactly once at create; never
  persisted. Lookup is by hash, so existence-vs-absence no longer leaks via
  row-scan timing. The middleware returns an opaque AuthenticationError for
  every failure mode so internal errors can't be fingerprinted.
- Idempotency rewritten claim-first. Repo replaces racy Exists+Create with a
  single Claim(key, ttl) -> (claimed, err) that uses ON CONFLICT DO NOTHING.
  Service claims BEFORE any side effect; releases on any downstream failure so
  the PSP's retry can re-run. Hash upgraded from MD5 to SHA-256 and namespaces
  by PSP. Engine StartWorkflow / SignalSubscriptionWorkflow errors now
  propagate (previously discarded).
- ORDER BY SQL injection closed. postgres.Paginate validates the sort column
  against ^[a-z_][a-z0-9_]*$ and the direction against {ASC,DESC}; bad input
  falls back to 'created_at DESC'. Limit clamped to [1,200].
- Cedar admin rule scoped to its own org. The previous unconditional
  permit(principal == Role::"admin", action, resource) was a platform-wide
  super-role; added when { principal.org_id == resource.org_id }.
- SSRF gate + timeout on outgoing webhooks. URLs are pre-resolved at Create AND
  at Send; any IP that isn't public-unicast (loopback/private/CGNAT/metadata)
  is rejected. The httpClient has a 30s total timeout and a DialContext that
  re-checks the resolved IP at connect time (DNS-rebinding defense). Response
  body is drained capped to 64KiB.
- PII / payment-token logging removed. Webhook bodies, Checkout.com request
  JSON, and paymentMethod.Token are no longer logged. Only routing fields
  (psp, size, reference) survive at INFO.

High (correctness):
- Subscription state machine wrapped in SELECT FOR UPDATE + tx. Activate /
  Pause / Resume / Cancel call FindByIdForUpdate inside TxManager.RunInTx so
  concurrent transitions can't both read the same initial state and overwrite
  each other.
- HTTP server timeouts set (ReadHeaderTimeout 5s, ReadTimeout 30s,
  WriteTimeout 60s, IdleTimeout 120s) — slowloris-resistant.
- DB pool configured (MaxOpenConns 25, MaxIdleConns 10, ConnMaxLifetime 5m,
  ConnMaxIdleTime 1m). Defaults were unlimited + stale.
- clientIP no longer trusts X-Forwarded-For / X-Real-IP unconditionally. New
  TRUSTED_PROXIES CIDR list gates header trust; with no proxies configured
  (the safe default) headers are ignored and RemoteAddr wins.
- Constructors no longer panic on pubsub Subscribe failure: SubscriptionService,
  SubscriptionEventBridge, ReportEventBridge, ReportService, CustomerService,
  DunningOrchestrationService, and WorkflowService now return (*X, error).
  Every pubsub handler is wrapped in defer recover() so a single handler panic
  no longer kills the receive loop. WORKFLOW_ENGINE validation is also a
  returned error instead of a panic.
- Repo gorm.ErrRecordNotFound translated to port.ErrNotFound at the postgres
  boundary so services don't import gorm.
- Webhook.go stale-err bug at "no parser registered" path fixed.

Medium/Low:
- Prorated credit math switched from float round-trip to integer (no fractional
  cent loss).
- cron.ScheduleTask uses the injected logger instead of fmt.Println.
- Test coverage: scopes (allowlist + clamping), cedar admin org scoping,
  trusted-proxy clientIP, Paystack + Checkout HMAC verifiers, webhook URL
  safety + DialContext predicate, idempotency Claim/Release semantics
  including the new release-on-failure paths.

Env additions: CHECKOUT_WEBHOOK_SECRET, API_KEY_PEPPER, TRUSTED_PROXIES.

CI already runs go vet + go test -race ./... on push/PR — the stale memory
note claiming otherwise was wrong and has been deleted.

+++ internal/core/domain/subscription.go
@@ -127,8 +127,13 @@ func (s *Subscription) CalculateProrationDetails(
 			return details
 		}
 
-		creditAmount := int(float64(s.Amount) * float64(daysRemaining) / float64(totalDays))
-		details.CreditAmount = creditAmount
+		// Stay in integer minor-units — a float round-trip would lose cents
+		// asymmetrically and produce off-by-one discrepancies at the cent
+		// boundary. Same truncation semantics as before (toward zero), which
+		// favors the merchant slightly; if business wants banker's rounding
+		// or ceiling, change here in one place.
+		creditAmount := (s.Amount * int64(daysRemaining)) / int64(totalDays)
+		details.CreditAmount = int(creditAmount)
 		details.DaysCredited = daysRemaining
 	}
